spark-1.x/src/main/java/cn/lhfei/spark/hive/udf/SimpleUDFExample.java
/*
 * Copyright 2010-2011 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package cn.lhfei.spark.hive.udf;

import org.apache.hadoop.hive.ql.exec.Description;
import org.apache.hadoop.hive.ql.exec.UDF;
import org.apache.hadoop.io.Text;

/**
 * @version 0.1
 *
 * @author Hefei Li
 *
 * @since Nov 3, 2015
 */

@Description(name = "SimpleUDFExample", value = "returns 'hello x', where x is whatever you give it (STRING)", extended = "SELECT simpleudfexample('world') from foo limit 1;")
class SimpleUDFExample extends UDF {

	public Text evaluate(Text input) {
		if (input == null)
			return null;
		return new Text("Hello " + input.toString());
	}
}
